feat(amazon): support ALBRequestCountPerTarget scaling policies (#10160)
* feat(amazon): support ALBRequestCountPerTarget scaling policies

The UI for TargetTracking scaling policies didn't support ALBRequestCountPerTarget. This was the only missing metric documented here: https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-scaling-target-tracking.html

This requires setting the resourceLabel parameter which is a combination of part of both the ALB and target group ARNs. This metric also only works with the Sum statistic. I found that the chart component hard coded average so that had to be fixed up as well.

* add target group label

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>