feat(webhook): add webhook.auditLoggingEnabled config property (#4840)
* fix(webhook): fix format string in response-too-big error message

* feat(webhook): add webhook.auditLoggedEnabled config property

which defaults to false.  When true, log information about each webhook request and response.

* refactor(webhook): use a literal for WebhookConfiguration's logger class

to remove the trailing bits from

com.netflix.spinnaker.orca.webhook.config.WebhookConfiguration$$EnhancerBySpringCGLIB$$2337d14d

* refactor(webhook): move audit logging after validation

so we only log a message for requests we actually send/responses we actually receive.
This seems more correct, but also paves the way to including the sizes in the log
messages.  With request/response size limits it's a bounded amount of work.

* feat(webhook): include header byte count in audit logging

* refactor(webhook): introduce WebhookConfiguration.getRequestBodyContentLength

to facilitate adding request body length to the audit log

* feat(webhook): include request body content length in the audit log

* feat(webhook): include response body content length in audit log