feat(core): add rootId to PipelineExecution
so it's cheap to find the root id of a pipeline execution.  Without it, it's possible to
look through triggers and find parents until the root, but this is easier.

It's implemented so that if a pipeline's root id is the same as its id, it's a top-level
pipeline.  Otherwise (root id != id), it's a child.