feature(kork-artifacts): Adds entity storage feature
This commit adds the entity storage feature which allows large objects
in the context to be stored by using the artifact store.

The entity storage uses the handler pattern which allows developers to
specify how they want to handle any key or object in the execution
context. There are two main handlers: the storage and expansion
handlers.

The handlers are injected into our (de)serializer hooks, and are added
to our "registry". The registry is just a serde modifier.

For example, when a map is about to be serialized or deserialized, it
will first call the modifier to construct the appropriate
(de)seriailizer, which is our new serde hooks.

These hooks house the handlers, and when the hook starts it's serde
process, it will ask the handlers whether or not it can handle the
object. If the handler can, the hook will then pass the object to the
handler, and the handler can mutate it as needed like the common case of
wanting to store large context objects. The same is true for the
expansion handlers but instead of storing, it expands (does not
necessarily need to expand).

To turn this feature on, add this to your spinnkaer configuration

```yaml
artifactStore:
  enabled: true
  entities:
    enabled: true
```

Additionally clouddriver requires one additional flag because it is the
ONLY service so far that needs to expand the context. So this needs to
also be added to the clouddriver.yaml
```yaml
artifactStore.entities.expand: true
```

In addition this commit adds:
* New exceptions to make use of Spring bubbling up and not overusing
  `RuntimeException`
* addresses that artifacts should always have some reference. This is
  now enforced with the @Nonnull annotation. Additional we will no
  longer serialize null values.
* Some helper functions to ArtifactTypes and adds a "CommonPrefixes"
  that can help identify artifact, e.g.  embedded vs remote.
* Artifact type based filtering
  So this change allows for filtering to happen at the artifact store
  level, and also by artifact type.

  ```yaml
  artifactStore:
    exclude:
      - type: "embedded/map/base64"
        value: "bad-application"
  ```

  This will look ONLY exclude `embedded/map/base64` and only for
  the application named `bad-application`. The value field is a regex so
  it does not need to be a simple as this example.

Signed-off-by: xibz <bjp@apple.com>