fix(header): add HeaderAuthRequestMatcher to RequestHeaderAuthenticationFilter
so it's clear that neither /error nor /auth/logout need authentication.  Otherwise, when
there's no X-SPINNAKER-USER header present in an incoming request,
RequestHeaderAuthenticationFilter throws an exception, and spring redirects to /error.
But then /error is supposed to render just fine even in the absence of the
X-SPINNAKER-USER header.  This change accomplishes that.