feat(header): new module for authenticating based on the X-SPINNAKER-USER header (#7109)
* feat(header): new module for authenticating based on the X-SPINNAKER-USER header

enabled when header.enabled is true (defaults to false)

* fix(header): use kork-security's User for UserDetails

since at least some code in gate (e.g. the /auth/user and /credentials/{account} endpoints) expect it.

* refactor(web/test): rename AuthControllerTest to HeaderAuthTest

since it's really about testing header authentication, not the auth controller.

* refactor(web/test): use tomcat in HeaderAuthTest

to pave the way to test more exceptional cases.

* test(web): demonstrate behavior of GET /auth/user using header authentication

with no X-SPINNAKER-USER header.  Eventually this may be processed as a valid request by
an anonymous user.  Currently, the RequestHeaderAuthenticationFilter bean is configured to
throw an exception.  The exception handling leaves something to be desired, since the
response is HTML instead of json.

* fix(header): add SpinnakerTomcatErrorValve

to use json response bodies for failed http responses.  This replaces the previous default
behavior in tomcat that generates html response bodies.  tomcat does provide
JsonErrorReportValve, but this gives us more control and makes it more likely that callers
can get consistent response bodies from the various exception handling paths in the code.
In other words SpinnakerTomcatErrorValve attempts to build responses like the ones that
spring + tomcat's GenericErrorController generates.

* test(web): demonstrate that error/exception handling isn't set up properly with header auth

* fix(header): add HeaderAuthRequestMatcher to RequestHeaderAuthenticationFilter

so it's clear that neither /error nor /auth/logout need authentication.  Otherwise, when
there's no X-SPINNAKER-USER header present in an incoming request,
RequestHeaderAuthenticationFilter throws an exception, and spring redirects to /error.
But then /error is supposed to render just fine even in the absence of the
X-SPINNAKER-USER header.  This change accomplishes that.

* test(web): add explicit test coverage for SpinnakerTomcatErrorValve

* feat(web): remove @Primary annotation from fiatService bean in GateConfig

to pave the way for tests to replace it.

This is meant to be a no-op.  Both FiatService beans: fiatService and fiatLoginService
still remain in the execution context.  Any references to FiatService beans that use some
other name will fail, but it doesn't appear any of those exist.

* test(web): enable fiat in HeaderAuthTest

and simulate a role to demonstrate that fiat is actually called.

* feat(web): add /auth/rawUser endpoint

so it's possible to see what's in the security context.  /auth/user explicitly calls fiat,
so it's not clear what roles are actually in the security context via that endpoint.

* test(header): demonstrate that csrf forbids requests

* fix(header): disable csrf in header auth

so e.g. POST requests have a chance of working.

Yes, it's possible to disable csrf manually in HeaderAuthConfigurerAdapter.configure.
Call AuthConfigure.configure since it disables csrf, as well as other things that are
likely important.  If nothing else, this makes header auth more consistent with other auth
mechanisms.

* feat(header): set allowedAccounts in User objects when using header authentication

to match what e.g. gate-oauth does.

* feat(header): log the user into fiat as part of header authentication