feat(core): add rootId to PipelineExecution (#7164)
* refactor(front50/test): make stubbing of ExecutionLauncher.start more specific

in DependentPipelineStarterSpec to pave the way for an additional rootId argument.

* refactor(front50/test): remove unnecessary mocking of ArtifactUtils

It's a spy, not a mock, and the tests pass without the mocking.

* refactor(front50/test): ensure that executionLauncher.start/fail were really called

* refactor(front50/test): assign the id properly in mock responses in DependentPipelineStarterSpec

* refactor(front50): convert DependentPipelineStartSpec from groovy to java

The addition of junit-jupiter-engine means that
UnusedPipelineDisablePollingNotificationAgentTest where it wasn't previously.

* refactor(core/test): rename ExecutionLauncherTest to ExecutionLauncherConfigurationTest

to pave the way for a simpler, non-spring ExecutionLauncherTest

* feat(core): add rootId to PipelineExecution

so it's cheap to find the root id of a pipeline execution.  Without it, it's possible to
look through triggers and find parents until the root, but this is easier.

It's implemented so that if a pipeline's root id is the same as its id, it's a top-level
pipeline.  Otherwise (root id != id), it's a child.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>