chore(orca): restore changes from https://github.com/spinnaker/orca/pull/4566 (#7167)
that were apparently lost in the monorepo migration.

Without the junit-jupiter-engine dependency in orca/gradle/groovy.gradle,
WebhookServiceTest doesn't run as part of

$ ./gradlew orca:orca-webhook:test

The addition of junit-bom to spinnaker-dependencies isn't a functional change.  spock-core
brings it in even before this change, but now it's explicit, and matches what's in the
[individual kork repo](https://github.com/spinnaker/kork/blob/a7b6fdea15ec2e98c80093947f194afb0ee552f3/spinnaker-dependencies/spinnaker-dependencies.gradle#L57).