feat(orca/webhook): add account property to webhook stages and introduce WebhookAccountProcessor interface (#7168)
* feat(web): Add an endpoint to fetch account details

only if the caller is authorized to use the account

GET /credentials/{accountName}/authorized

* feat(webhook): add new account property to webhook stages

and a new configuration property webhook.requireAccount (default false).  If
webhook.requireAccount is true and the account is not present (i.e. a non-empty string),
the stage fails.

Beyond that the account property isn't used.

* feat(webhook): add webhook.validateAccount

which defaults to false.  When true, and an account is present, call clouddriver's GET
/credentials/authorized/{account} endpoint to validate that the current user has
permission to use the account.  If the user doesn't have permission, assume clouddriver's
response causes orca to throw an exception.

* test(webhook): add customHeaders to WebhookServiceTest

to pave the way for the addition of WebhookAccountProcessor, which changes the way headers
are determined.

* feat(webhook): call WebhookAccountProcessor.getHeaders in WebhookService.getRestTemplateData

if there's a WebhookAccountProcessor bean in the spring context.

This makes it possible to customize the http request headers sent in webhook stages.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>