feat(orca/webhook): add regex pattern matching to the webhook allow list (#7171)
* refactor(webhook): add a matchStrategy property to WebhookProperties.AllowedRequest

that currently has only one possible value: STARTS_WITH.

This isn't a behavior change, nor does it add any new functionality.  It paves the way to
add additional match strategies though.

* refactor(webhook/test): fix typo in test name

* refactor(webhook/test): make matchStrategy explicit in WebhookServiceTest

since we're testing behavior of an explicit strategy, not the default.

Also add a test to demonstrate behavior of an empty urlPrefix

* feat(webhook): add regex pattern matching to the webhook allow list

via a new value for webhook.allowedRequests[].matchStrategy: PATTERN_MATCHES

When matchStrategy is PATTERN_MATCHES, a URL is allowed if it matches webhook.allowedRequests[].urlPattern.

* feat(webhook): add a more specific error message when urlPrefix is null

when using the STARTS_WITH strategy.

* refactor(webhook): validate allowedRequest configuration at startup

so an invalid configuration causes orca to fail to start, instead of starting and failing when executing a webhook stage.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>