feat(webhook): add new webhook allow list property safeToRetry (#7172)
* refactor(webhook): make RestTemplateData a top-level class

to pave the way for using it outside WebhookService

* refactor(webhook): make RestTemplateData available to CreateWebhookTask.execute

to pave the way for using RestTemplateData in WebhookResponseProcessor.

* refactor(webhook): add Optional<AllowedRequest> to RestTemplateData

so it's available when processing responses.

* feat(webhook): add new webhook allow list property safeToRetry

which defaults to false.

Before this commit, CreateWebhookTask would retry SocketTimeoutException only for GET
requests.  Now, it also retries other request methods (e.g. PUT, POST, etc.) if
safeToRetry is true in the relevant allow list entry.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>