feat(orca): plumb additional headers from incoming requests into webhook stages (#7179)
* feat(web): add additionalHeaders to SpinnakerRequestHeaderInterceptor

that are propagated from the MDC to outgoing http requests using retrofit2, when the value
in the MDC is non-empty.

Configured via ok-http-client.additioanlHeaders which defaults to an empty list.

* feat(web): move ProvidedIdRequestFilter from gate

so it's available to other Spinnaker microservices (e.g. orca).

In combination with additionalHeaders in SpinnakerRequestHeaderInterceptor, this makes it
possible to pass headers from incoming http requests to gate into orca's MDC.

* test(web): demonstrate behavior of WebConfiguration

to pave the way for verifying upcoming bean creation logic there.

* feat(web): include ProvidedIdRequestFilter in the spring context

AuthenticatedRequestFilter already puts incoming X-SPINNAKER-* headers in the MDC.
ProvidedIdRequestFilter has an additionalHeaders property which makes it interesting to
use as well.

Provide a ProvidedIdRequestFilterConfigurationProperties bean as well via
EnableConfigurationProperties.

* feat(core): add new PipelineExecution property additionalHeaders

If provided-id-request-filter.enabled is true, populate the PipelineExecution's
additionalHeaders with the provided-id-request-filter.additional-headers that have
non-empty values from the http request that initiated the pipeline execution.

* feat(queue): add additional headers from the execution context into the MDC of running tasks

Add logback-test to enable the use of MemoryAppender in tests.  Without it, there's no
appender in the root logger.

* feat(webhook): include additional headers from ProvidedIdRequestFilter in webhook calls

when the new config flag webhook.includeAdditionalHeaders is true (default false), and the
ProvidedIdRequestFilter is enabled.

Any additional headers that ProvidedIdRequestFilter is configured to include are dropped
from customHeaders, whether they're present or not.  This allows API providers to trust the
values of the configured additional headers.