feat(clouddriver/aws): Implement the account API for aws accounts (#7238)
* test(clouddriver/aws): Demonstrate behavior of account API for AWS accounts

Specifically, that the account API doesn't work for AWS accounts (yet).

* feat(clouddriver/aws): Implement the account API for aws accounts

This change allows clouddriver AWS accounts to use the account management
API. Specifically, it allows clouddriver to load AWS accounts from the
AccountDefinitionRepository used by the account API.

The new config flag account.storage.aws.enabled, in addition to the existing config flag
account.storage.enabled, control when this feature is enabled.

* fix(account): Provide defaults to avoid NPEs

An account without requiredGroupMembership or permissions is a
valid account. However, certain areas of the code assume that these
values are not null and therefore encounter NullPointerExceptions.
To avoid this, initialize these fields to a reasonable default value.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>