fix(sql): Update account type column during upserts (#7242)
* test(account): Define behavior when replacing an account

Replacing an account ("accountOne")  with a new account ("accountTwo")
of the same name but different type should replace all of accountOne's
data with accountTwo's data.

* fix(sql): Update account type column during upserts

If a user creates an account with an initial type, say "aws", and
then creates another account with the same name but a different
type, say "kubernetes", the operation will succeed but the type
given by the "type" column will not match the actual type of the
account. This causes clouddriver to fail to start as it cannot
marshal the account definition jsons into the correct Java objects.
Fix this by also updating the "type" column during an update
operation.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>