Various kork updates (secrets, utilities, prereqs for credentials updates) (#7245)
* update(kotlin): Upgrade to Kotlin 1.8

* chore(editorconfig): Combine editorconfig files

* feature(kork-security): Support spring security principals

* chore(kork-secrets): Remove unused file

* refactor(kork-core): Extract ClassScanner utility

* feature(kork-core): Add NamedTypeProvider SPI

* refactor(kork-core): Extract BootstrapComponents

* feature(kork-secrets-api): add `boot-secret://` properties

This extracts `kork-security-api` from `kork-security` with some porting to Kotlin to help reduce some code duplication. This allows for replacing secrets at startup in configuration data.

* fix(clouddriver-core): Fix use of secrets API

* fix(keiko-core): Handle all QueueEvent types

Newer versions of Kotlin require that all types of a sealed type be handled in `when` expressions, and the compiler found a couple missing event types in `QueueMetricsPublisher`.

* fix(test): Fix secrets-related tests from clouddriver

API updates to secrets require some updates to clouddriver tests that will be moved to kork in a later update.

* fix(keel-scm): Fix use of kotlin-reflect

* fix(kork-plugins): Fix compiler errors for Kotlin 1.8

* fix(orca-kayenta): Stop using experimental APIs

* fix(keel): Update mockk to support Kotlin 1.8

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>