fix(orca): set the status correctly when cancelling a pipeline execution (#7296)
* test(orca/core): demonstrate how status is set when canceling a pipeline

and specifically, that canceling a running pipeline with one completed stage and one
not-started stage doesn't actually set the status of the overall execution to canceled.

* test(orca/core): add a stage to pipelines used for cancel testing

to pave the way for updates to the logic for setting pipeline status during cancel operations.

* fix(orca): set the status correctly when cancelling a pipeline execution

The exact details to reproduce this race are unfortunately lost to time.  We've been
running this code at Salesforce since April 2021 though, so it's got some good miles on
it.  It also still seems like a reasonable fix.

---------

Co-authored-by: Arifullah Pattan <apattan@salesforce.com>