perf(execution): Add a retrieve method with requireLatestVersion (#7310)
* perf(execution): Add a retrieve method with requireLatestVersion

Adds a retrieve method with the requireLatestVersion option. This option
is useful in use cases where the execution repository has a read
replica and retrieve()'s caller does not need the latest version of the
PipelineExecution. In these cases, the caller can retrieve the
PipelineExecution currently stored in the read replica. This is a
much cheaper operation than querying the main writer instance.

* perf(repository): Use a more efficient db query

for polled endpoints. Deck periodically polls the /pipelines and
/pipelines/{id} endpoints, so orca doesn't need to provide
the latest pipeline execution info. Specifically for databases
with read replicas, this means that orca will return the pipeline
execution(s) stored in the read replica without needing to deal
with replication lag.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>
Co-authored-by: Jason <mcintoshj@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>