perf(gate/web): increase performance of the GET /{application}/pipelineConfigs/{pipelineName:.+} endpoint (#7314)
* refactor(gate/web/test): include exception handling in ApplicationControllerSpec

to pave the way for changing some of the underlying front50 endpoints

* test(gate/web): demonstrate behavior of GET /{application}/pipelineConfigs/{pipelineName:.+}

when front50 responds with 500

* refactor(gate/web/test): remove mock web server since it's not used

* refactor(gate/web/test): refer to individual pipelines by name in ApplicationControllerSpec

to reduce differences when changing the front50 endpoint

* refactor(gate/core): add type arguments to raw Map in Front50Service

Specifically, getPipelineConfigsForApplication and getStrategyConfigs to facilitate
type-safe mocking in tests, and to generally improve code quality.

* refactor(gate/web/test): convert ApplicationControllerSpec from groovy to java

for general goodness, but also to pave the way to convert from standaloneSetup to
webAppContextSetup where mixing spock mocks and spring mock beans doesn't end well.

* refactor(orca/web/test): convert ApplicationControllerTest from standaloneSetup to webAppContextSetup

to pave the way for additional exception handling.  standaloneSetup only accepts one
object to use for ControllerAdvice, but with exception handling in both
GenericExceptionHandlers and SpinnakerRetrofitExceptionHandlers, that isn't sufficient.

* feat(gate/web): include the application name in the exception message from ApplicationService.getPipelineConfigForApplication

so the message is more helpful for callers who don't augment it.

* perf(gate/web): increase performance of the GET /{application}/pipelineConfigs/{pipelineName:.+} endpoint

by changing it to query front50 for a specific pipeline instead of all pipelines in an
application, and then filtering.

* fix(gate/core): remove leading slashes in retrofit URL

to minimize diffs, and to match what https://github.com/spinnaker/spinnaker/pull/7159 did.