feat(gate/executions): add a pipelineNameFilter query parameter when getting pipeline executions (#7315)
* refactor(web/test): reduce duplication in TaskControllerSqlExecutionRepositoryTest

Refactoring the TaskControllerSqlExecutionRepositoryTest class to avoid some copy pasta.

This change also modifies the endpoints being tested to use
/v2/applications/{application}/pipelines instead of the normal
/applications/{application}/pipelines. The v2 version of the endpoint simply passes
through the parameters to unversioned endpoint, and the v2 version is the one that
actually gets used, so it's better to call into the v2 version for better test coverage.

* perf(orca/executions): add pipelineNameFilter to /applications/{application}/pipelines endpoint

and /v2/applications/{application}/pipelines endpoint as well.

This adds the pipelineNameFilter to the endpoints which get pipelines by application
name. Right now this filter is present in gate, but gate does the filtering with the
parameter, instead of passing it through to orca. This is inefficient as there is no need
for orca to send over all of the pipeline execution information for an application, and
then have gate do the filtering. With this functionality here, we can change gate to
simply pass the parameter through.

* feat(gate/executions): add a pipelineNameFilter when getting pipeline executions

This adds a pipelineNameFilter to the /applications/{application}/pipelines endpoint. This
provides functionality to filter pipeline executions by pipeline name. This paves the way for a UI optimization.

* perf(gate/executions): pass pipelineNameFilter to orca

This change passes the pipelineNameFilter to orca.

With this change the /application/{applications}/pipelines endpoint should get a lot of
optimization, because we will avoid passing all of the pipeline execution data from orca
to gate when it is not needed.

---------

Co-authored-by: Richard Timpson <richard.timpson@salesforce.com>