perf(executions): add pipelineLimit query parameter to gate's GET /applications/{application}/pipelines and GET /applications/{application}/pipelineConfigs endpoints (#7319)
* fix(front50/pipelines): fix bug rewriting pipeline index in cache

Fixes a bug that rewrites the pipeline index value of the pipelines in the front50 cache. The previous code that set p.setIndex(i) was actually rewriting the index values of the pipelines in the cache, because the p is actually a reference to the underlying memory of the cache. To fix, we do a shallow clone of the pipeline objects that we are going to return, and set the index on the cloned objects.

This bug became apparent with the pipelineLimit query parameter, as subsequent calls to the same endpoint with the same limit value were returning different results. This also adds a specific test case for that very scenario, to ensure that front50 always uses the index values that exist in the db.

* perf(front50/pipelines): pipelineLimit query param on /pipelineConfigs

Adds a pipelineLimit query param on the /{application}/pipelineConfigs endpoint. This parameter ensures that the number of pipelines returned from the endpoint never exceeds the number specified by pipelineLimit.

Combined with similar changes to gate and orca, this paves the way for a UI change to limit the number of pipeline configs and executions on the executions page.

* perf(orca/pipelines): pipelineLimit query param on /pipelines

Adds a pipelineLimit to the /v2/applications/{application}/pipelines endpoint. The endpoint passes through this query parameter to front50 when fetching pipeline configs. Because the endpoint only uses the pipeline configs ids that it gets from front50, this gives us a mechanism to limit the number of executions returned.

* refactor(orca/pipelines): remove unnecessary conditional logic when fetching pipeline executions

Small change that removes an unnecessary ternary expression when calling front50 to get pipeline configs while fetching executions by application.

It's more straightforward to always pass through pipelineLimit and pipelineNameFilter. If they are null, the params will simply be left out of the request to front50.

* perf(gate/pipelines): new pipelineLimit param when fetching configs

Pass through the pipelineLimit query param to front50 when fetching pipelineConfigs. This allows the UI to control how many configs are returned in an attempt to optimize the executions page for larger applications.

* perf(gate/executions): new pipelineLimit param when fetching executions

Pass through the pipelineLimit query param to orca when fetching pipeline executions.  This allows the UI to control how many executions are returned when displaying a list of executions per application.

---------

Co-authored-by: Richard Timpson <richard.timpson@salesforce.com>