fix(orca/sql): Retrieve the newest pipeline executions (#7320)
* refactor(orca/web/test): use newer assertions library

* fix(orca/sql): Retrieve the newest pipeline executions

for a particular pipeline config_id

Apologies for not having a test here.  We noticed this after migrating live instances of
Spinnaker to mysql 8.  We didn't see the issue with mysql 5.7, and our automated tests
were running with mysql 5.7 at the time.

Without this fix, we would see newer executions disappear from the UI, replaced by older
ones.

---------

Co-authored-by: Daniel Zheng <d.zheng@salesforce.com>