refactor(orca/sql): add selectFields + DSLContext.selectExecution functions to SqlExecutionRepository
These are identical to the ones in jooq.kt, but they don't need a compressionProperties argument since it's a member of SqlExecutionRepository.

Instead of implementing them to call the functions in jooq, leave the duplication to
reduce conflicts with other changes coming in.  Once those are complete, we can revisit
reducing the duplication if it remains.